add game resources check (length and sha) add game splash check (length and sha) add game archive check (length and sha) new game resources store algorithm

+++ InAppStorySdk/inappstorysdk/src/main/java/com/inappstory/sdk/stories/api/models/GameSplashScreen.java
@@ -6,7 +6,7 @@ public class GameSplashScreen {
     @SerializedName("url")
     public String url;
     @SerializedName("size")
-    public int size;
+    public Long size;
     @SerializedName("sha1")
     public String sha1;
 }
